requirements.txt
transformers
sentencepiece
Pillow
matplotlib
protobuf